requirements.txt
pytest==5.4.1
opencv_python==4.2.0.32
Flask==1.1.1
six==1.14.0
requests==2.23.0
numpy==1.18.4
tensorflow_gpu==2.1.0
Pillow==7.1.2
tensorflow==2.2.0
